FoodIngSubModel/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.11.0
python-dotenv==1.0.0
openai==1.13.3
supabase==2.4.1
requests==2.31.0